app-93 update scenatio
